backend/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "arkeep-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "PyJWT>=2.8",
    "google-auth>=2.29",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
